Data Scientists/Data Analyst/ Statistician Job Tracker

Job advertisements

DSraw

Format

A data frame with 551 rows and 152 variables

ID

row id

Consultant

Name of the consultant

DateRetrieved

Date of Data Retrieved

DatePublished

Published Date of the Advertisement

Job_title

Name of the job category

Company

Name of the Company

R

If R is required -> 1 ,If not mentioned -> 0

SAS

If SAS is required -> 1 , If not mentioned -> 0

SPSS

If SPSS is required -> 1 , If not mentioned -> 0

Python

If Python is required -> 1 , If not mentioned -> 0

MAtlab

If Matlab is required -> 1 , If not mentioned -> 0

Scala

If Scala is required -> 1 , If not mentioned -> 0

C#

If C# is required -> 1 , If not mentioned -> 0

MS Word

If knowledge in MS Word is required -> 1 , If not mentioned -> 0

Ms Excel

If knowledge in MS Excel is required -> 1 , If not mentioned -> 0

OLE/DB

If knowledge in OLE/DB is required -> 1 , If not mentioned -> 0

Ms Access

If Ms Access is required -> 1 , If not mentioned -> 0

Ms PowerPoint

If knowledge in Ms Powerpoint is required -> 1 , If not mentioned -> 0

Spreadsheets

If knowledge in Spreadsheets is required -> 1 , If not mentioned -> 0

Data_visualization

If knowledge inData Visualization is required -> 1 , If not mentioned -> 0

Presentation_Skills

If Presentation Skills are required -> 1 , If not mentioned -> 0

Communication

If Communication skills are required -> 1 , If not mentioned -> 0

BigData

If knowledge in Big Data analysis is required -> 1 , If not mentioned -> 0

Data_warehouse

If knowledge in Data Warehouse is required -> 1 , If not mentioned -> 0

cloud_storage

If knowledge in Cloud Storage is required -> 1 , If not mentioned -> 0

Google_Cloud

If knowledge in Google Cloud is required -> 1 , If not mentioned -> 0

AWS

If knowledge in AWS is required -> 1 , If not mentioned -> 0

Machine_Learning

If knowledge in Machine Learning is required -> 1 , If not mentioned -> 0

Deep Learning

If knowledge in Deep Learning is required -> 1 , If not mentioned -> 0

Computer_vision

If knowledge in Computer Vision is required -> 1 , If not #' mentioned -> 0

Java

If Java is required -> 1 , If not mentioned -> 0

C++

If C++ is required -> 1 , If not mentioned -> 0

C

If C is required -> 1 , If not mentioned -> 0

Linux/Unix

If knowledge in Linux/Unix is required -> 1 , If not mentioned -> 0

SQL

If SQL is required -> 1 , If not mentioned -> 0

NoSQL

If NoSQL is required -> 1 , If not mentioned -> 0

RDBMS

If knowledge in RDBMS is required -> 1 , If not mentioned -> 0

Oracle

If knowledge in Oracle is required -> 1 , If not mentioned -> 0

MySQL

If MYSQL is required -> 1 , If not mentioned -> 0

PHP

If PHP is required -> 1 , If not mentioned -> 0

Flash_Actionscript

If knowledge in Flash Action Script is required -> 1 , If not mentioned -> 0

SPL

If knowledge in SPL is required -> 1 , If not mentioned -> 0

web_design_and_development_tools

If knowledge in Web Design and Development Tools is required -> 1 , If not mentioned -> 0

Wordpress

If knowledge in Wordpress is required -> 1 , If not mentioned -> 0

AI

If Artificial Intelligence is required -> 1 , If not mentioned -> 0

Natural_Language_Processing(NLP)

If knowledge in NLP is required -> 1 , If not mentioned -> 0

Microsoft Power BI

If knowledge in Microsoft Power BI is required -> 1 , If not mentioned -> 0

Google_Analytics

If knowledge in Google Analytics is required -> 1 , If not mentioned -> 0

graphics_and_design_skills

If Graphic and Design Skills are required -> 1 , If not mentioned -> 0

Data_marketing

If Data Marketing abillity is required -> 1 , If not mentioned -> 0

SEO

If knowledge in SEO is required -> 1 , If not mentioned -> 0

Content_Management

If knowledge in Content Management is required -> 1 , If not mentioned -> 0

Tableau

If knowledge in Tableau is required -> 1 , If not mentioned -> 0

D3

If knowledge in D3 is required -> 1 , If not mentioned -> 0

Alteryx

If knowledge in Alteryx is required -> 1 , If not mentioned -> 0

KNIME

If knowledge in KNIME is required -> 1 , If not mentioned -> 0

Spotfire

If knowledge in Spotfire is required -> 1 , If not mentioned -> 0

Spark

If knowledge in Spark is required -> 1 , If not mentioned -> 0

S3

If knowledge in S3 is required -> 1 , If not mentioned -> 0

Redshift

If knowledge in Redshift is required -> 1 , If not mentioned -> 0

DigitalOcean

If knowledge in Digital Ocean is required -> 1 , If not mentioned -> 0

Javascript

If Java Script is required -> 1 , If not mentioned -> 0

Kafka

If knowledge in Kafka is required -> 1 , If not mentioned -> 0

Storm

If knowledge in Storm is required -> 1 , If not mentioned -> 0

Bash

If knowledge in Bash is required -> 1 , If not mentioned -> 0

Hadoop

If knowledge in Hadoop is required -> 1 , If not mentioned -> 0

Data_Pipelines

If knowledge in Data Pipelines is required -> 1 , If not mentioned -> 0

MPP_Platforms

If MPP Platforms is required ->1,If not mentioned-0

Qlik

If Qlik is required ->1,If not mentioned ->0

Pig

If Pig is required ->1,If not mentioned ->0

Hive

If Hive is required ->1,If not mentioned ->0

Tensorflow

If Tensorflow is required ->1,If not mentioned ->0

Map/Reduce

If Map/Reduce is required ->1,If not mentioned ->0

Impala

If Impala is required ->1,If not mentioned ->0

Solr

If Sloris required ->1,If not mentioned ->0

Teradata

If Teradata is required ->1,If not mentioned ->0

MongoDB

If MonoDB is required ->1,If not mentioned ->0

Elasticsearch

If Elasticsearch is required ->1,If not mentioned ->0

YOLO

If YOLO is required-1 ,If not mentioned-0

agile execution

If agile execution is required->1 ,If not mentioned->0

Data_management

If the knowledge in data management is required->1 ,If not mentioned->0

pyspark

If pyspark is required->1 ,If not mentioned->0

Data_mining

If the knowledge in data mining is required->1 ,If not mentioned->0

Data_science

If the knowledge in data science is required->1 ,If not mentioned->0

Web_Analytic_tools

If the knowledge in Web Analytic tools is required->1 ,If not mentioned->0

IOT

If IOT is required->1 ,If not mentioned->0

Numerical_Analysis

If the knowledge in Numerical Analysis is required->1 ,If not mentioned->0

Economic

If the knowledge in Economic is required->1 ,If not mentioned->0

Finance_Knowledge

If Finance_Knowledge is required->1 ,If not mentioned->0

Investment_Knowledge

If Investment Knowledge is required->1 ,If not mentioned->0

Problem_Solving

If the ability of Problem Solving is required->1 ,If not mentioned->0

Korean_language

If the ability of speaking Korean language is required->1 ,If not mentioned->0

Bash\Linux Scripting

If Bash\ Linux Scripting is required->1 ,If not mentioned->0

Knowledge_in

Required knowledge to do a particular job ,If not mentioned->NA

Experience

Minimum experience required for a particular job

City

City where the company is located in

Location

Country where the company is located in

Educational_qualifications

Required educational qualifications

Salary

Amount of salary

Team_Handling

If the ability of Team Handling is required-1 ,If not mentioned-0

Debtor_reconcilation

If the ability of Debtor reconciliation is required-1 ,If not mentioned-0

Payroll_management

If the ability of Payroll management is required-1 ,If not mentioned-0

Bayesian

If Bayesian knowledge is required-1 ,If not mentioned-0

Optimization

If Optimization knowledge is required-1 ,If not mentioned-0

Bahasa Malaysia

If Bahasa Malaysia is required-1 ,If not mentioned-0

English proficiency

If English proficiency is required-1 ,If not mentioned-0

URL

Web address of a particular job advertisement

Search_Term

web search term of a particular job advertisement

X109

Columns with null values

X110

Columns with null values

X111

Columns with null values

X112

Columns with null values

X113

Columns with null values

X114

Columns with null values

X115

Columns with null values

X116

Columns with null values

X117

Columns with null values

X118

Columns with null values

X119

Columns with null values

X120

Columns with null values

X121

Columns with null values

X122

Columns with null values

X123

Columns with null values

X124

Columns with null values

X125

Columns with null values

X126

Columns with null values

X127

Columns with null values

X128

Columns with null values

X129

Columns with null values

X130

Columns with null values

X131

Columns with null values

X132

Columns with null values

X133

Columns with null values

X134

Columns with null values

X135

Columns with null values

X136

Columns with null values

X137

Columns with null values

X138

Columns with null values

X139

Columns with null values

X140

Columns with null values

X141

Columns with null values

X142

Columns with null values

X143

Columns with null values

X144

Columns with null values

X145

Columns with null values

X146

Columns with null values

X147

Columns with null values

X148

Columns with null values

X149

Columns with null values

X150

Columns with null values

X151

Columns with null values

X152

Columns with null values

Source

Collected and entered by BSc (Hons) Statistics undegraduates - 2020

Examples

data(DSraw)
head(DSraw)

#> # A tibble: 6 x 152
#>      ID Consultant DateRetrieved DatePublished Job_title Company     R   SAS
#>   <dbl> <chr>      <chr>         <chr>         <chr>     <chr>   <dbl> <dbl>
#> 1     1 Thiyanga   05/08/2020    NA            NA        NA          1     1
#> 2     2 Jayani     07/08/2020    31/07/2020    Junior D… Dialog…     1     0
#> 3     3 Jayani     07/08/2020    06/08/20      Engineer… London…     0     0
#> 4     4 Jayani     07/08/2020    24/07/2020    CI-Stati… E.D. B…     1     1
#> 5     5 Jayani     07/08/2020    24/07/2020    DA-Data … E.D. B…     0     1
#> 6     6 Jayani     07/08/2020    13/08/2020    Data Sci… Emirat…     1     0
#> # … with 144 more variables: SPSS <dbl>, Python <dbl>, MAtlab <dbl>,
#> #   Scala <dbl>, `C#` <dbl>, `MS Word` <dbl>, `Ms Excel` <dbl>, `OLE/DB` <dbl>,
#> #   `Ms Access` <dbl>, `Ms PowerPoint` <dbl>, Spreadsheets <dbl>,
#> #   Data_visualization <dbl>, Presentation_Skills <dbl>, Communication <dbl>,
#> #   BigData <dbl>, Data_warehouse <dbl>, cloud_storage <dbl>,
#> #   Google_Cloud <dbl>, AWS <dbl>, Machine_Learning <dbl>, `Deep
#> #   Learning` <dbl>, Computer_vision <dbl>, Java <dbl>, `C++` <dbl>, C <dbl>,
#> #   `Linux/Unix` <dbl>, SQL <dbl>, NoSQL <dbl>, RDBMS <dbl>, Oracle <dbl>,
#> #   MySQL <dbl>, PHP <dbl>, Flash_Actionscript <dbl>, SPL <dbl>,
#> #   web_design_and_development_tools <dbl>, Wordpress <dbl>, AI <dbl>,
#> #   `Natural_Language_Processing(NLP)` <dbl>, `Microsoft Power BI` <dbl>,
#> #   Google_Analytics <dbl>, graphics_and_design_skills <dbl>,
#> #   Data_marketing <dbl>, SEO <dbl>, Content_Management <dbl>, Tableau <dbl>,
#> #   D3 <dbl>, Alteryx <dbl>, KNIME <dbl>, Spotfire <dbl>, Spark <dbl>,
#> #   S3 <dbl>, Redshift <dbl>, DigitalOcean <dbl>, Javascript <dbl>,
#> #   Kafka <dbl>, Storm <dbl>, Bash <dbl>, Hadoop <dbl>, Data_Pipelines <dbl>,
#> #   MPP_Platforms <dbl>, Qlik <dbl>, Pig <dbl>, Hive <dbl>, Tensorflow <dbl>,
#> #   `Map/Reduce` <dbl>, Impala <dbl>, Solr <dbl>, Teradata <dbl>,
#> #   MongoDB <dbl>, Elasticsearch <dbl>, YOLO <dbl>, `agile execution` <dbl>,
#> #   Data_management <dbl>, pyspark <dbl>, Data_mining <dbl>,
#> #   Data_science <dbl>, Web_Analytic_tools <dbl>, IOT <dbl>,
#> #   Numerical_Analysis <dbl>, Economic <dbl>, Finance_Knowledge <dbl>,
#> #   Investment_Knowledge <dbl>, Problem_Solving <dbl>, Korean_language <dbl>,
#> #   `Bash\\Linux Scripting` <dbl>, Knowledge_in <chr>, Experience <chr>,
#> #   City <chr>, Location <chr>, Educational_qualifications <chr>, Salary <chr>,
#> #   Team_Handling <dbl>, Debtor_reconcilation <dbl>, Payroll_management <dbl>,
#> #   Bayesian <dbl>, Optimization <dbl>, `Bahasa Malaysia` <dbl>, `English
#> #   proficiency` <chr>, URL <chr>, Search_Term <chr>, …

summary(DSraw)

#>        ID         Consultant        DateRetrieved      DatePublished     
#>  Min.   :  1.0   Length:551         Length:551         Length:551        
#>  1st Qu.:138.2   Class :character   Class :character   Class :character  
#>  Median :275.5   Mode  :character   Mode  :character   Mode  :character  
#>  Mean   :275.4                                                           
#>  3rd Qu.:412.8                                                           
#>  Max.   :550.0                                                           
#>  NA's   :1                                                               
#>   Job_title           Company                R               SAS        
#>  Length:551         Length:551         Min.   :0.0000   Min.   :0.0000  
#>  Class :character   Class :character   1st Qu.:0.0000   1st Qu.:0.0000  
#>  Mode  :character   Mode  :character   Median :1.0000   Median :0.0000  
#>                                        Mean   :0.5207   Mean   :0.2135  
#>                                        3rd Qu.:1.0000   3rd Qu.:0.0000  
#>                                        Max.   :1.0000   Max.   :1.0000  
#>                                        NA's   :140      NA's   :167     
#>       SPSS             Python           MAtlab            Scala        
#>  Min.   :0.00000   Min.   :0.0000   Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.00000   1st Qu.:0.0000   1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.00000   Median :1.0000   Median :0.00000   Median :0.00000  
#>  Mean   :0.09574   Mean   :0.6745   Mean   :0.08136   Mean   :0.07902  
#>  3rd Qu.:0.00000   3rd Qu.:1.0000   3rd Qu.:0.00000   3rd Qu.:0.00000  
#>  Max.   :1.00000   Max.   :1.0000   Max.   :1.00000   Max.   :1.00000  
#>  NA's   :175       NA's   :127      NA's   :170       NA's   :184      
#>        C#             MS Word        Ms Excel         OLE/DB       
#>  Min.   :0.00000   Min.   :0.00   Min.   :0.000   Min.   :0.00000  
#>  1st Qu.:0.00000   1st Qu.:0.00   1st Qu.:0.000   1st Qu.:0.00000  
#>  Median :0.00000   Median :0.00   Median :0.000   Median :0.00000  
#>  Mean   :0.00552   Mean   :0.08   Mean   :0.221   Mean   :0.00552  
#>  3rd Qu.:0.00000   3rd Qu.:0.00   3rd Qu.:0.000   3rd Qu.:0.00000  
#>  Max.   :1.00000   Max.   :1.00   Max.   :1.000   Max.   :1.00000  
#>  NA's   :189       NA's   :176    NA's   :171     NA's   :189      
#>    Ms Access      Ms PowerPoint      Spreadsheets     Data_visualization
#>  Min.   :0.0000   Min.   :0.00000   Min.   :0.00000   Min.   :0.0000    
#>  1st Qu.:0.0000   1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.0000    
#>  Median :0.0000   Median :0.00000   Median :0.00000   Median :0.0000    
#>  Mean   :0.0615   Mean   :0.09019   Mean   :0.01381   Mean   :0.4314    
#>  3rd Qu.:0.0000   3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:1.0000    
#>  Max.   :1.0000   Max.   :1.00000   Max.   :1.00000   Max.   :1.0000    
#>  NA's   :177      NA's   :174       NA's   :189       NA's   :150       
#>  Presentation_Skills Communication       BigData       Data_warehouse  
#>  Min.   :0.0000      Min.   :0.0000   Min.   :0.0000   Min.   :0.0000  
#>  1st Qu.:0.0000      1st Qu.:0.0000   1st Qu.:0.0000   1st Qu.:0.0000  
#>  Median :0.0000      Median :1.0000   Median :0.0000   Median :0.0000  
#>  Mean   :0.2104      Mean   :0.5396   Mean   :0.3258   Mean   :0.1037  
#>  3rd Qu.:0.0000      3rd Qu.:1.0000   3rd Qu.:1.0000   3rd Qu.:0.0000  
#>  Max.   :1.0000      Max.   :1.0000   Max.   :1.0000   Max.   :1.0000  
#>  NA's   :185         NA's   :147      NA's   :155      NA's   :175     
#>  cloud_storage     Google_Cloud          AWS          Machine_Learning
#>  Min.   :0.0000   Min.   :0.00000   Min.   :0.00000   Min.   :0.0000  
#>  1st Qu.:0.0000   1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.0000  
#>  Median :0.0000   Median :0.00000   Median :0.00000   Median :0.0000  
#>  Mean   :0.1671   Mean   :0.01377   Mean   :0.04972   Mean   :0.4702  
#>  3rd Qu.:0.0000   3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:1.0000  
#>  Max.   :1.0000   Max.   :1.00000   Max.   :1.00000   Max.   :1.0000  
#>  NA's   :162      NA's   :188       NA's   :189       NA's   :132     
#>  Deep Learning    Computer_vision       Java             C++         
#>  Min.   :0.0000   Min.   :0.0000   Min.   :0.0000   Min.   :0.00000  
#>  1st Qu.:0.0000   1st Qu.:0.0000   1st Qu.:0.0000   1st Qu.:0.00000  
#>  Median :0.0000   Median :0.0000   Median :0.0000   Median :0.00000  
#>  Mean   :0.1303   Mean   :0.0551   Mean   :0.1175   Mean   :0.05249  
#>  3rd Qu.:0.0000   3rd Qu.:0.0000   3rd Qu.:0.0000   3rd Qu.:0.00000  
#>  Max.   :1.0000   Max.   :1.0000   Max.   :1.0000   Max.   :1.00000  
#>  NA's   :175      NA's   :188      NA's   :185      NA's   :189      
#>        C             Linux/Unix            SQL             NoSQL        
#>  Min.   :0.00000   Min.   : 0.00000   Min.   :0.0000   Min.   :0.00000  
#>  1st Qu.:0.00000   1st Qu.: 0.00000   1st Qu.:0.0000   1st Qu.:0.00000  
#>  Median :0.00000   Median : 0.00000   Median :1.0000   Median :0.00000  
#>  Mean   :0.02747   Mean   : 0.05525   Mean   :0.5931   Mean   :0.06011  
#>  3rd Qu.:0.00000   3rd Qu.: 0.00000   3rd Qu.:1.0000   3rd Qu.:0.00000  
#>  Max.   :1.00000   Max.   :10.00000   Max.   :1.0000   Max.   :1.00000  
#>  NA's   :187       NA's   :189        NA's   :143      NA's   :185      
#>      RDBMS             Oracle          MySQL            PHP         
#>  Min.   :0.00000   Min.   :0.000   Min.   :0.000   Min.   :0.00000  
#>  1st Qu.:0.00000   1st Qu.:0.000   1st Qu.:0.000   1st Qu.:0.00000  
#>  Median :0.00000   Median :0.000   Median :0.000   Median :0.00000  
#>  Mean   :0.02747   Mean   :0.064   Mean   :0.115   Mean   :0.01337  
#>  3rd Qu.:0.00000   3rd Qu.:0.000   3rd Qu.:0.000   3rd Qu.:0.00000  
#>  Max.   :1.00000   Max.   :1.000   Max.   :1.000   Max.   :1.00000  
#>  NA's   :187       NA's   :176     NA's   :177     NA's   :177      
#>  Flash_Actionscript      SPL          web_design_and_development_tools
#>  Min.   :0.00000    Min.   :0.00000   Min.   :0.00000                 
#>  1st Qu.:0.00000    1st Qu.:0.00000   1st Qu.:0.00000                 
#>  Median :0.00000    Median :0.00000   Median :0.00000                 
#>  Mean   :0.00535    Mean   :0.01604   Mean   :0.01867                 
#>  3rd Qu.:0.00000    3rd Qu.:0.00000   3rd Qu.:0.00000                 
#>  Max.   :1.00000    Max.   :1.00000   Max.   :1.00000                 
#>  NA's   :177        NA's   :177       NA's   :176                     
#>    Wordpress         AI         Natural_Language_Processing(NLP)
#>  Min.   :0     Min.   :0.0000   Min.   :0.0000                  
#>  1st Qu.:0     1st Qu.:0.0000   1st Qu.:0.0000                  
#>  Median :0     Median :0.0000   Median :0.0000                  
#>  Mean   :0     Mean   :0.1326   Mean   :0.1392                  
#>  3rd Qu.:0     3rd Qu.:0.0000   3rd Qu.:0.0000                  
#>  Max.   :0     Max.   :1.0000   Max.   :1.0000                  
#>  NA's   :177   NA's   :159      NA's   :163                     
#>  Microsoft Power BI Google_Analytics  graphics_and_design_skills
#>  Min.   :0.0000     Min.   :0.00000   Min.   :0.00000           
#>  1st Qu.:0.0000     1st Qu.:0.00000   1st Qu.:0.00000           
#>  Median :0.0000     Median :0.00000   Median :0.00000           
#>  Mean   :0.1223     Mean   :0.01102   Mean   :0.01877           
#>  3rd Qu.:0.0000     3rd Qu.:0.00000   3rd Qu.:0.00000           
#>  Max.   :1.0000     Max.   :1.00000   Max.   :1.00000           
#>  NA's   :175        NA's   :188       NA's   :178               
#>  Data_marketing         SEO          Content_Management    Tableau   
#>  Min.   :0.00000   Min.   :0.00000   Min.   :0.00000    Min.   :0.0  
#>  1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.00000    1st Qu.:0.0  
#>  Median :0.00000   Median :0.00000   Median :0.00000    Median :0.0  
#>  Mean   :0.00829   Mean   :0.00276   Mean   :0.00276    Mean   :0.2  
#>  3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:0.00000    3rd Qu.:0.0  
#>  Max.   :1.00000   Max.   :1.00000   Max.   :1.00000    Max.   :1.0  
#>  NA's   :189       NA's   :189       NA's   :189        NA's   :181  
#>        D3             Alteryx            KNIME            Spotfire      
#>  Min.   :0.00000   Min.   :0.00000   Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.00000   Median :0.00000   Median :0.00000   Median :0.00000  
#>  Mean   :0.02486   Mean   :0.01934   Mean   :0.00552   Mean   :0.01108  
#>  3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:0.00000  
#>  Max.   :1.00000   Max.   :1.00000   Max.   :1.00000   Max.   :1.00000  
#>  NA's   :189       NA's   :189       NA's   :189       NA's   :190      
#>      Spark              S3             Redshift        DigitalOcean    
#>  Min.   :0.0000   Min.   :0.00000   Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.0000   1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.0000   Median :0.00000   Median :0.00000   Median :0.00000  
#>  Mean   :0.1078   Mean   :0.01108   Mean   :0.01934   Mean   :0.00829  
#>  3rd Qu.:0.0000   3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:0.00000  
#>  Max.   :1.0000   Max.   :1.00000   Max.   :1.00000   Max.   :1.00000  
#>  NA's   :180      NA's   :190       NA's   :189       NA's   :189      
#>    Javascript          Kafka            Storm              Bash        
#>  Min.   :0.00000   Min.   :0.0000   Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.00000   1st Qu.:0.0000   1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.00000   Median :0.0000   Median :0.00000   Median :0.00000  
#>  Mean   :0.03867   Mean   :0.0221   Mean   :0.00276   Mean   :0.00276  
#>  3rd Qu.:0.00000   3rd Qu.:0.0000   3rd Qu.:0.00000   3rd Qu.:0.00000  
#>  Max.   :1.00000   Max.   :1.0000   Max.   :1.00000   Max.   :1.00000  
#>  NA's   :189       NA's   :189      NA's   :189       NA's   :189      
#>      Hadoop       Data_Pipelines    MPP_Platforms          Qlik        
#>  Min.   :0.0000   Min.   :0.00000   Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.0000   1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.0000   Median :0.00000   Median :0.00000   Median :0.00000  
#>  Mean   :0.1162   Mean   :0.02486   Mean   :0.00276   Mean   :0.04959  
#>  3rd Qu.:0.0000   3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:0.00000  
#>  Max.   :1.0000   Max.   :1.00000   Max.   :1.00000   Max.   :1.00000  
#>  NA's   :181      NA's   :189       NA's   :189       NA's   :188      
#>       Pig               Hive           Tensorflow        Map/Reduce     
#>  Min.   :0.00000   Min.   :0.00000   Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.00000   Median :0.00000   Median :0.00000   Median :0.00000  
#>  Mean   :0.01657   Mean   :0.07609   Mean   :0.04408   Mean   :0.01105  
#>  3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:0.00000  
#>  Max.   :1.00000   Max.   :1.00000   Max.   :1.00000   Max.   :1.00000  
#>  NA's   :189       NA's   :183       NA's   :188       NA's   :189      
#>      Impala             Solr            Teradata          MongoDB       
#>  Min.   :0.00000   Min.   :0.00000   Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.00000   Median :0.00000   Median :0.00000   Median :0.00000  
#>  Mean   :0.02198   Mean   :0.00276   Mean   :0.00826   Mean   :0.01377  
#>  3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:0.00000  
#>  Max.   :1.00000   Max.   :1.00000   Max.   :1.00000   Max.   :1.00000  
#>  NA's   :187       NA's   :189       NA's   :188       NA's   :188      
#>  Elasticsearch          YOLO         agile execution   Data_management  
#>  Min.   :0.00000   Min.   :0.00000   Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.00000   Median :0.00000   Median :0.00000   Median :0.00000  
#>  Mean   :0.01381   Mean   :0.00276   Mean   :0.02762   Mean   :0.01934  
#>  3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:0.00000   3rd Qu.:0.00000  
#>  Max.   :1.00000   Max.   :1.00000   Max.   :1.00000   Max.   :1.00000  
#>  NA's   :189       NA's   :189       NA's   :189       NA's   :189      
#>     pyspark         Data_mining     Data_science    Web_Analytic_tools
#>  Min.   :0.00000   Min.   :0.000   Min.   :0.0000   Min.   :0.00000   
#>  1st Qu.:0.00000   1st Qu.:0.000   1st Qu.:0.0000   1st Qu.:0.00000   
#>  Median :0.00000   Median :0.000   Median :0.0000   Median :0.00000   
#>  Mean   :0.01939   Mean   :0.159   Mean   :0.3333   Mean   :0.01105   
#>  3rd Qu.:0.00000   3rd Qu.:0.000   3rd Qu.:1.0000   3rd Qu.:0.00000   
#>  Max.   :1.00000   Max.   :1.000   Max.   :1.0000   Max.   :1.00000   
#>  NA's   :190       NA's   :180     NA's   :188      NA's   :189       
#>       IOT          Numerical_Analysis    Economic       Finance_Knowledge
#>  Min.   :0.00000   Min.   :0.00000    Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.00000   1st Qu.:0.00000    1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.00000   Median :0.00000    Median :0.00000   Median :0.00000  
#>  Mean   :0.00826   Mean   :0.03867    Mean   :0.05525   Mean   :0.04144  
#>  3rd Qu.:0.00000   3rd Qu.:0.00000    3rd Qu.:0.00000   3rd Qu.:0.00000  
#>  Max.   :1.00000   Max.   :1.00000    Max.   :1.00000   Max.   :1.00000  
#>  NA's   :188       NA's   :189        NA's   :189       NA's   :189      
#>  Investment_Knowledge Problem_Solving  Korean_language   Bash\\Linux Scripting
#>  Min.   :0.00000      Min.   :0.0000   Min.   :0.00000   Min.   :0.00000      
#>  1st Qu.:0.00000      1st Qu.:0.0000   1st Qu.:0.00000   1st Qu.:0.00000      
#>  Median :0.00000      Median :0.0000   Median :0.00000   Median :0.00000      
#>  Mean   :0.00552      Mean   :0.2707   Mean   :0.00276   Mean   :0.01374      
#>  3rd Qu.:0.00000      3rd Qu.:1.0000   3rd Qu.:0.00000   3rd Qu.:0.00000      
#>  Max.   :1.00000      Max.   :1.0000   Max.   :1.00000   Max.   :1.00000      
#>  NA's   :189          NA's   :189      NA's   :189       NA's   :187          
#>  Knowledge_in        Experience            City             Location        
#>  Length:551         Length:551         Length:551         Length:551        
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  Educational_qualifications    Salary          Team_Handling   
#>  Length:551                 Length:551         Min.   :0.0000  
#>  Class :character           Class :character   1st Qu.:0.0000  
#>  Mode  :character           Mode  :character   Median :0.0000  
#>                                                Mean   :0.3609  
#>                                                3rd Qu.:1.0000  
#>                                                Max.   :1.0000  
#>                                                NA's   :188     
#>  Debtor_reconcilation Payroll_management    Bayesian        Optimization    
#>  Min.   :0.00000      Min.   :0.00000    Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.00000      1st Qu.:0.00000    1st Qu.:0.00000   1st Qu.:0.00000  
#>  Median :0.00000      Median :0.00000    Median :0.00000   Median :0.00000  
#>  Mean   :0.00276      Mean   :0.00276    Mean   :0.01102   Mean   :0.08287  
#>  3rd Qu.:0.00000      3rd Qu.:0.00000    3rd Qu.:0.00000   3rd Qu.:0.00000  
#>  Max.   :1.00000      Max.   :1.00000    Max.   :1.00000   Max.   :1.00000  
#>  NA's   :189          NA's   :189        NA's   :188       NA's   :189      
#>  Bahasa Malaysia   English proficiency     URL            Search_Term       
#>  Min.   :0.00000   Length:551          Length:551         Length:551        
#>  1st Qu.:0.00000   Class :character    Class :character   Class :character  
#>  Median :0.00000   Mode  :character    Mode  :character   Mode  :character  
#>  Mean   :0.00276                                                            
#>  3rd Qu.:0.00000                                                            
#>  Max.   :1.00000                                                            
#>  NA's   :189                                                                
#>    X109           X110           X111           X112           X113        
#>  Mode:logical   Mode:logical   Mode:logical   Mode:logical   Mode:logical  
#>  NA's:551       NA's:551       NA's:551       NA's:551       NA's:551      
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>    X114           X115           X116           X117           X118        
#>  Mode:logical   Mode:logical   Mode:logical   Mode:logical   Mode:logical  
#>  NA's:551       NA's:551       NA's:551       NA's:551       NA's:551      
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>    X119           X120           X121           X122           X123        
#>  Mode:logical   Mode:logical   Mode:logical   Mode:logical   Mode:logical  
#>  NA's:551       NA's:551       NA's:551       NA's:551       NA's:551      
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>    X124           X125           X126           X127           X128        
#>  Mode:logical   Mode:logical   Mode:logical   Mode:logical   Mode:logical  
#>  NA's:551       NA's:551       NA's:551       NA's:551       NA's:551      
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>    X129           X130           X131           X132           X133        
#>  Mode:logical   Mode:logical   Mode:logical   Mode:logical   Mode:logical  
#>  NA's:551       NA's:551       NA's:551       NA's:551       NA's:551      
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>    X134           X135           X136           X137           X138        
#>  Mode:logical   Mode:logical   Mode:logical   Mode:logical   Mode:logical  
#>  NA's:551       NA's:551       NA's:551       NA's:551       NA's:551      
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>    X139           X140           X141           X142           X143        
#>  Mode:logical   Mode:logical   Mode:logical   Mode:logical   Mode:logical  
#>  NA's:551       NA's:551       NA's:551       NA's:551       NA's:551      
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>    X144           X145           X146           X147           X148        
#>  Mode:logical   Mode:logical   Mode:logical   Mode:logical   Mode:logical  
#>  NA's:551       NA's:551       NA's:551       NA's:551       NA's:551      
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>    X149           X150           X151           X152        
#>  Mode:logical   Mode:logical   Mode:logical   Mode:logical  
#>  NA's:551       NA's:551       NA's:551       NA's:551      
#>                                                             
#>                                                             
#>                                                             
#>                                                             
#>