Data Scientists/Data Analyst/ Statistician Job Tracker

Job advertisements

DStidy

Format

A data frame with 430 rows and 114 variables

ID

row id

Consultant

Name of the consultant

DateRetrieved

Date of Data Retrieved

DatePublished

Published Date of the Advertisement

Job_title

Name of the job category

Company

Name of the Company

R

If R is required -> 1 ,If not mentioned -> 0

SAS

If SAS is required -> 1 , If not mentioned -> 0

SPSS

If SPSS is required -> 1 , If not mentioned -> 0

Python

If Python is required -> 1 , If not mentioned -> 0

MAtlab

If Matlab is required -> 1 , If not mentioned -> 0

Scala

If Scala is required -> 1 , If not mentioned -> 0

C_Sharp

If C# is required -> 1 , If not mentioned -> 0

MS_Word

If knowledge in MS Word is required -> 1 , If not mentioned -> 0

Ms_Excel

If knowledge in MS Excel is required -> 1 , If not mentioned -> 0

OLE_DB

If knowledge in OLE/DB is required -> 1 , If not mentioned -> 0

Ms_Access

If Ms Access is required -> 1 , If not mentioned -> 0

Ms_PowerPoint

If knowledge in Ms Powerpoint is required -> 1 , If not mentioned -> 0

Spreadsheets

If knowledge in Spreadsheets is required -> 1 , If not mentioned -> 0

Data_visualization

If knowledge inData Visualization is required -> 1 , If not mentioned -> 0

Presentation_Skills

If Presentation Skills are required -> 1 , If not mentioned -> 0

Communication

If Communication skills are required -> 1 , If not mentioned -> 0

BigData

If knowledge in Big Data analysis is required -> 1 , If not mentioned -> 0

Data_warehouse

If knowledge in Data Warehouse is required -> 1 , If not mentioned -> 0

cloud_storage

If knowledge in Cloud Storage is required -> 1 , If not mentioned -> 0

Google_Cloud

If knowledge in Google Cloud is required -> 1 , If not mentioned -> 0

AWS

If knowledge in AWS is required -> 1 , If not mentioned -> 0

Machine_Learning

If knowledge in Machine Learning is required -> 1 , If not mentioned -> 0

Deep_Learning

If knowledge in Deep Learning is required -> 1 , If not entioned -> 0

Computer_vision

If knowledge in Computer Vision is required -> 1 , If not mentioned -> 0

Java

If Java is required -> 1 , If not mentioned -> 0

Cpp

If C++ is required -> 1 , If not mentioned -> 0

C

If C is required -> 1 , If not mentioned -> 0

Linux_Unix

If knowledge in Linux/Unix is required -> 1 , If not mentioned -> 0

SQL

If SQL is required -> 1 , If not mentioned -> 0

NoSQL

If NoSQL is required -> 1 , If not mentioned -> 0

RDBMS

If knowledge in RDBMS is required -> 1 , If not mentioned -> 0

Oracle

If knowledge in Oracle is required -> 1 , If not mentioned -> 0

MySQL

If MYSQL is required -> 1 , If not mentioned -> 0

PHP

If PHP is required -> 1 , If not mentioned -> 0

Flash_Actionscript

If knowledge in Flash Action Script is required -> 1 , If not mentioned -> 0

SPL

If knowledge in SPL is required -> 1 , If not mentioned -> 0

web_design_and_development_tools

If knowledge in Web Design and Development Tools is required -> 1 , If not mentioned -> 0

Wordpress

If knowledge in Wordpress is required -> 1 , If not mentioned -> 0

AI

If Artificial Intelligence is required -> 1 , If not mentioned -> 0

Natural_Language_Processing(NLP)

If knowledge in NLP is required -> 1 , If not mentioned -> 0

Microsoft_Power_BI

If knowledge in Microsoft Power BI is required -> 1 , If not mentioned -> 0

Google_Analytics

If knowledge in Google Analytics is required -> 1 , If not mentioned -> 0

graphics_and_design_skills

If Graphic and Design Skills are required -> 1 , If not mentioned -> 0

Data_marketing

If Data Marketing abillity is required -> 1 , If not mentioned -> 0

SEO

If knowledge in SEO is required -> 1 , If not mentioned -> 0

Content_Management

If knowledge in Content Management is required -> 1 , If not mentioned -> 0

Tableau

If knowledge in Tableau is required -> 1 , If not mentioned -> 0

D3

If knowledge in D3 is required -> 1 , If not mentioned -> 0

Alteryx

If knowledge in Alteryx is required -> 1 , If not mentioned -> 0

KNIME

If knowledge in KNIME is required -> 1 , If not mentioned -> 0

Spotfire

If knowledge in Spotfire is required -> 1 , If not mentioned -> 0

Spark

If knowledge in Spark is required -> 1 , If not mentioned -> 0

S3

If knowledge in S3 is required -> 1 , If not mentioned -> 0

Redshift

If knowledge in Redshift is required -> 1 , If not mentioned -> 0

DigitalOcean

If knowledge in Digital Ocean is required -> 1 , If not mentioned -> 0

Javascript

If Java Script is required -> 1 , If not mentioned -> 0

Kafka

If knowledge in Kafka is required -> 1 , If not mentioned -> 0

Storm

If knowledge in Storm is required -> 1 , If not mentioned -> 0

Bash

If knowledge in Bash is required -> 1 , If not mentioned -> 0

Hadoop

If knowledge in Hadoop is required -> 1 , If not mentioned -> 0

Data_Pipelines

If knowledge in Data Pipelines is required -> 1 , If not mentioned -> 0

MPP_Platforms

If MPP Platforms is required ->1,If not mentioned-0

Qlik

If Qlik is required ->1,If not mentioned ->0

Pig

If Pig is required ->1,If not mentioned ->0

Hive

If Hive is required ->1,If not mentioned ->0

Tensorflow

If Tensorflow is required ->1,If not mentioned ->0

Map_Reduce

If Map/Reduce is required ->1,If not mentioned ->0

Impala

If Impala is required ->1,If not mentioned ->0

Solr

If Sloris required ->1,If not mentioned ->0

Teradata

If Teradata is required ->1,If not mentioned ->0

MongoDB

If MonoDB is required ->1,If not mentioned ->0

Elasticsearch

If Elasticsearch is required ->1,If not mentioned ->0

YOLO

If YOLO is required-1 ,If not mentioned-0

agile_execution

If agile execution is required->1 ,If not mentioned->0

Data_management

If the knowledge in data management is required->1 ,If not mentioned->0

pyspark

If pyspark is required->1 ,If not mentioned->0

Data_mining

If the knowledge in data mining is required->1 ,If not mentioned->0

Data_science

If the knowledge in data science is required->1 ,If not mentioned->0

Web_Analytic_tools

If the knowledge in Web Analytic tools is required->1 ,If not mentioned->0

IOT

If IOT is required->1 ,If not mentioned->0

Numerical_Analysis

If the knowledge in Numerical Analysis is required->1 ,If not mentioned->0

Economic

If the knowledge in Economic is required->1 ,If not mentioned->0

Finance_Knowledge

If Finance_Knowledge is required->1 ,If not mentioned->0

Investment_Knowledge

If Investment Knowledge is required->1 ,If not mentioned->0

Problem_Solving

If the ability of Problem Solving is required->1 ,If not mentioned->0

Korean_language

If the ability of speaking Korean language is required->1 ,If not mentioned->0

Bash_Linux_Scripting

If Bash\ Linux Scripting is required->1 ,If not mentioned->0

Knowledge_in

Required knowledge to do a particular job ,If not mentioned->NA

City

City where the company is located in

Location

Country where the company is located in

Educational_qualifications

Required educational qualifications

Salary

Amount of salary

Team_Handling

If the ability of Team Handling is required-1 ,If not mentioned-0

Debtor_reconcilation

If the ability of Debtor reconciliation is required-1 ,If not mentioned-0

Payroll_management

If the ability of Payroll management is required-1 ,If not mentioned-0

Bayesian

If Bayesian knowledge is required-1 ,If not mentioned-0

Optimization

If Optimization knowledge is required-1 ,If not mentioned-0

Bahasa_Malaysia

If Bahasa Malaysia is required-1 ,If not mentioned-0

English_proficiency

If English proficiency is required-1 ,If not mentioned-0

URL

Web address of a particular job advertisement

Search_Term

web search term of a particular job advertisement

Job_Category

Category of the job (i.e. "Data Science","Data Analyst" etc.)

Minimum_Years_of_experience

Minimum years of experience needed for the job

Experience

Minimum experience required for a particular job

Experience_Category

Number of years of experience in binned into categories

Job_Country

Country where job is located in

Edu_Category

Type of educational qualification needed

Minimum_Salary

Starting salary proposed in respective currency

Salary_Basis

Payment basis of salary (i.e. "unspecified","hourly","daily","monthly","yearly")

Source

Data wrangling part was done by Chammika R. De Mel, Shashini D. Silva and Thimani D Ranathunga, BSc (Hons)Staistics, University of Sri Jayewardenepura.

Examples

data(DStidy)
head(DStidy)

#> # A tibble: 6 x 115
#>      ID Consultant DateRetrieved DatePublished Job_title Company R     SAS  
#>   <int> <chr>      <date>        <date>        <chr>     <chr>   <fct> <fct>
#> 1     1 Thiyanga   2020-08-05    NA            NA        NA      1     1    
#> 2     2 Jayani     2020-08-07    2020-07-31    Junior D… Dialog… 1     0    
#> 3     3 Jayani     2020-08-07    2020-08-06    Engineer… London… 0     0    
#> 4     4 Jayani     2020-08-07    2020-07-24    CI-Stati… E.D. B… 1     1    
#> 5     5 Jayani     2020-08-07    2020-07-24    DA-Data … E.D. B… 0     1    
#> 6     6 Jayani     2020-08-07    2020-08-13    Data Sci… Emirat… 1     0    
#> # … with 107 more variables: SPSS <fct>, Python <fct>, MAtlab <fct>,
#> #   Scala <fct>, C_Sharp <fct>, MS_Word <fct>, Ms_Excel <fct>, OLE_DB <fct>,
#> #   Ms_Access <fct>, Ms_PowerPoint <fct>, Spreadsheets <fct>,
#> #   Data_visualization <fct>, Presentation_Skills <fct>, Communication <fct>,
#> #   BigData <fct>, Data_warehouse <fct>, cloud_storage <fct>,
#> #   Google_Cloud <fct>, AWS <fct>, Machine_Learning <fct>, Deep_Learning <fct>,
#> #   Computer_vision <fct>, Java <fct>, Cpp <fct>, C <fct>, Linux_Unix <fct>,
#> #   SQL <fct>, NoSQL <fct>, RDBMS <fct>, Oracle <fct>, MySQL <fct>, PHP <fct>,
#> #   Flash_Actionscript <fct>, SPL <fct>,
#> #   web_design_and_development_tools <fct>, Wordpress <fct>, AI <fct>,
#> #   `Natural_Language_Processing(NLP)` <fct>, Microsoft_Power_BI <fct>,
#> #   Google_Analytics <fct>, graphics_and_design_skills <fct>,
#> #   Data_marketing <fct>, SEO <fct>, Content_Management <fct>, Tableau <fct>,
#> #   D3 <fct>, Alteryx <fct>, KNIME <fct>, Spotfire <fct>, Spark <fct>,
#> #   S3 <fct>, Redshift <fct>, DigitalOcean <fct>, Javascript <fct>,
#> #   Kafka <fct>, Storm <fct>, Bash <fct>, Hadoop <fct>, Data_Pipelines <fct>,
#> #   MPP_Platforms <fct>, Qlik <fct>, Pig <fct>, Hive <fct>, Tensorflow <fct>,
#> #   Map_Reduce <fct>, Impala <fct>, Solr <fct>, Teradata <fct>, MongoDB <fct>,
#> #   Elasticsearch <fct>, YOLO <fct>, agile_execution <fct>,
#> #   Data_management <fct>, pyspark <fct>, Data_mining <fct>,
#> #   Data_science <fct>, Web_Analytic_tools <fct>, IOT <fct>,
#> #   Numerical_Analysis <fct>, Economic <fct>, Finance_Knowledge <fct>,
#> #   Investment_Knowledge <fct>, Problem_Solving <fct>, Korean_language <fct>,
#> #   Bash_Linux_Scripting <fct>, Team_Handling <fct>,
#> #   Debtor_reconcilation <fct>, Payroll_management <fct>, Bayesian <fct>,
#> #   Optimization <fct>, Bahasa_Malaysia <fct>, Knowledge_in <chr>, City <chr>,
#> #   Location <chr>, Educational_qualifications <chr>, Salary <chr>,
#> #   English_proficiency <chr>, URL <chr>, Search_Term <chr>,
#> #   Job_Category <fct>, …

summary(DStidy)

#>        ID         Consultant        DateRetrieved        DatePublished       
#>  Min.   :  1.0   Length:430         Min.   :2020-05-13   Min.   :2018-07-04  
#>  1st Qu.:108.2   Class :character   1st Qu.:2020-08-10   1st Qu.:2020-07-24  
#>  Median :232.5   Mode  :character   Median :2020-08-13   Median :2020-08-05  
#>  Mean   :250.8                      Mean   :2020-08-16   Mean   :2020-07-24  
#>  3rd Qu.:383.8                      3rd Qu.:2020-08-13   3rd Qu.:2020-08-11  
#>  Max.   :536.0                      Max.   :2020-10-16   Max.   :2020-08-28  
#>                                     NA's   :7            NA's   :6           
#>   Job_title           Company          R       SAS     SPSS    Python  MAtlab 
#>  Length:430         Length:430         0:220   0:349   0:394   0:148   0:399  
#>  Class :character   Class :character   1:210   1: 81   1: 36   1:282   1: 31  
#>  Mode  :character   Mode  :character                                          
#>                                                                               
#>                                                                               
#>                                                                               
#>                                                                               
#>  Scala   C_Sharp MS_Word Ms_Excel OLE_DB  Ms_Access Ms_PowerPoint Spreadsheets
#>  0:401   0:428   0:400   0:347    0:428   0:408     0:396         0:425       
#>  1: 29   1:  2   1: 30   1: 83    1:  2   1: 22     1: 34         1:  5       
#>                                                                               
#>                                                                               
#>                                                                               
#>                                                                               
#>                                                                               
#>  Data_visualization Presentation_Skills Communication BigData Data_warehouse
#>  0:260              0:353               0:214         0:304   0:391         
#>  1:170              1: 77               1:216         1:126   1: 39         
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  cloud_storage Google_Cloud AWS     Machine_Learning Deep_Learning
#>  0:365         0:425        0:412   0:235            0:381        
#>  1: 65         1:  5        1: 18   1:195            1: 49        
#>                                                                   
#>                                                                   
#>                                                                   
#>                                                                   
#>                                                                   
#>  Computer_vision Java    Cpp     C       Linux_Unix SQL     NoSQL   RDBMS  
#>  0:410           0:387   0:411   0:420   0:419      0:191   0:408   0:420  
#>  1: 20           1: 43   1: 19   1: 10   1: 11      1:239   1: 22   1: 10  
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>  Oracle  MySQL   PHP     Flash_Actionscript SPL    
#>  0:406   0:389   0:426   0:428              0:424  
#>  1: 24   1: 41   1:  4   1:  2              1:  6  
#>                                                    
#>                                                    
#>                                                    
#>                                                    
#>                                                    
#>  web_design_and_development_tools Wordpress AI     
#>  0:423                            0:430     0:378  
#>  1:  7                                      1: 52  
#>                                                    
#>                                                    
#>                                                    
#>                                                    
#>                                                    
#>  Natural_Language_Processing(NLP) Microsoft_Power_BI Google_Analytics
#>  0:378                            0:386              0:426           
#>  1: 52                            1: 44              1:  4           
#>                                                                      
#>                                                                      
#>                                                                      
#>                                                                      
#>                                                                      
#>  graphics_and_design_skills Data_marketing SEO     Content_Management Tableau
#>  0:423                      0:427          0:429   0:429              0:357  
#>  1:  7                      1:  3          1:  1   1:  1              1: 73  
#>                                                                              
#>                                                                              
#>                                                                              
#>                                                                              
#>                                                                              
#>  D3      Alteryx KNIME   Spotfire Spark   S3      Redshift DigitalOcean
#>  0:421   0:423   0:428   0:426    0:390   0:426   0:423    0:427       
#>  1:  9   1:  7   1:  2   1:  4    1: 40   1:  4   1:  7    1:  3       
#>                                                                        
#>                                                                        
#>                                                                        
#>                                                                        
#>                                                                        
#>  Javascript Kafka   Storm   Bash    Hadoop  Data_Pipelines MPP_Platforms
#>  0:416      0:422   0:429   0:429   0:387   0:421          0:429        
#>  1: 14      1:  8   1:  1   1:  1   1: 43   1:  9          1:  1        
#>                                                                         
#>                                                                         
#>                                                                         
#>                                                                         
#>                                                                         
#>  Qlik    Pig     Hive    Tensorflow Map_Reduce Impala  Solr    Teradata MongoDB
#>  0:412   0:424   0:402   0:414      0:426      0:422   0:429   0:427    0:425  
#>  1: 18   1:  6   1: 28   1: 16      1:  4      1:  8   1:  1   1:  3    1:  5  
#>                                                                                
#>                                                                                
#>                                                                                
#>                                                                                
#>                                                                                
#>  Elasticsearch YOLO    agile_execution Data_management pyspark Data_mining
#>  0:425         0:429   0:421           0:423           0:423   0:372      
#>  1:  5         1:  1   1:  9           1:  7           1:  7   1: 58      
#>                                                                           
#>                                                                           
#>                                                                           
#>                                                                           
#>                                                                           
#>  Data_science Web_Analytic_tools IOT     Numerical_Analysis Economic
#>  0:311        0:426              0:427   0:417              0:410   
#>  1:119        1:  4              1:  3   1: 13              1: 20   
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>                                                                     
#>  Finance_Knowledge Investment_Knowledge Problem_Solving Korean_language
#>  0:415             0:428                0:334           0:429          
#>  1: 15             1:  2                1: 96           1:  1          
#>                                                                        
#>                                                                        
#>                                                                        
#>                                                                        
#>                                                                        
#>  Bash_Linux_Scripting Team_Handling Debtor_reconcilation Payroll_management
#>  0:425                0:303         0:429                0:429             
#>  1:  5                1:127         1:  1                1:  1             
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>                                                                            
#>  Bayesian Optimization Bahasa_Malaysia Knowledge_in           City          
#>  0:426    0:401        0:429           Length:430         Length:430        
#>  1:  4    1: 29        1:  1           Class :character   Class :character  
#>                                        Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>    Location         Educational_qualifications    Salary         
#>  Length:430         Length:430                 Length:430        
#>  Class :character   Class :character           Class :character  
#>  Mode  :character   Mode  :character           Mode  :character  
#>                                                                  
#>                                                                  
#>                                                                  
#>                                                                  
#>  English_proficiency     URL            Search_Term       
#>  Length:430          Length:430         Length:430        
#>  Class :character    Class :character   Class :character  
#>  Mode  :character    Mode  :character   Mode  :character  
#>                                                           
#>                                                           
#>                                                           
#>                                                           
#>                             Job_Category Minimum_Years_of_experience
#>  Data Analyst                     : 23   Min.   : 0.000             
#>  Data Science                     :210   1st Qu.: 1.000             
#>  Data Science and Data Engineering:  8   Median : 3.000             
#>  Unimportant                      :189   Mean   : 3.027             
#>                                          3rd Qu.: 4.750             
#>                                          Max.   :13.000             
#>                                          NA's   :100                
#>   Experience                                Experience_Category
#>  Length:430         Unknown or Not needed             :100     
#>  Class :character   Two or less years                 :162     
#>  Mode  :character   More than 2 and less than 5 years :125     
#>                     More than 5 and less than 10 years: 42     
#>                     More than 10 years                :  1     
#>                                                                
#>                                                                
#>  Job_Country             Edu_Category Minimum_Salary           Salary_Basis
#>  Length:430         Other      : 25   Min.   :      0.0   daily      :  1  
#>  Class :character   Some Degree: 90   1st Qu.:     27.2   hourly     :  3  
#>  Mode  :character   Min_Bsc    :147   Median :     84.0   unspecified:409  
#>                     Min_Master : 58   Mean   :  88148.8   yearly     : 17  
#>                     Phd        : 17   3rd Qu.:  12965.0                    
#>                     NA's       : 93   Max.   :1500000.0                    
#>                                       NA's   :366