Data scientists, data Analyst, and statistician related job advertisements in 2020

A dataset with 430 rows and 115 columns

data(DStidy_2020)

Source

Data wrangling part was done by Janith C. Wanniarachchie, BSc (Hons)Staistics, University of Sri Jayewardenepura and description file was prepared by Randi Shashikala.

Details

ID. Row id

Consultant. Name of the consultant

DateRetrieved. Date of data retrieved

DatePublished. Published date of the advertisement

Job_title. Name of the job category

Company. Name of the company

R. If R is required -> 1 , If not mentioned -> 0

SAS. If SAS is required -> 1 , If not mentioned -> 0

SPSS. If SPSS is required -> 1 , If not mentioned -> 0

Python. If Python is required -> 1 , If not mentioned -> 0

MAtlab. If MAtlab is required -> 1 , If not mentioned -> 0

Scala. If Scala is required -> 1 , If not mentioned -> 0

C_Sharp. If C_Sharp is required -> 1 , If not mentioned -> 0

Ms_Excel. If Ms_Excel is required -> 1 , If not mentioned -> 0

OLE_DB. If OLE_DB is required -> 1 , If not mentioned -> 0

Ms_Access. If Ms_Access is required -> 1 , If not mentioned -> 0

Ms_PowerPoint. If Ms_PowerPoint is required -> 1 , If not mentioned -> 0

Spreadsheets. If Spreadsheets is required -> 1 , If not mentioned -> 0

Data_visualization. If knowledge in Data Visualization is required -> 1 , If not mentioned -> 0

Presentation_Skills. If Presentation Skills are required -> 1 , If not mentioned -> 0

Communication. If Communication skills are required -> 1 , If not mentioned -> 0

BigData. If knowledge in Big Data analysis is required -> 1 , If not mentioned -> 0

Data_warehouse. If knowledge in Data Warehouse is required -> 1 , If not mentioned -> 0

cloud_storage. If knowledge in Cloud Storage is required -> 1 , If not mentioned -> 0

Google_Cloud. If knowledge in Google Cloud is required -> 1 , If not mentioned -> 0

AWS. If knowledge in AWS is required -> 1 , If not mentioned -> 0

Machine_Learning. If knowledge in Machine Learning is required -> 1 , If not mentioned -> 0

Deep_Learning. If knowledge in Deep Learning is required -> 1 , If not mentioned -> 0

Computer_vision. If knowledge in Computer Vision is required -> 1 , If not mentioned -> 0

Java. If Java is required -> 1 , If not mentioned -> 0

Cpp. If Cpp is required -> 1 , If not mentioned -> 0

C. If C is required -> 1 , If not mentioned -> 0

Linux_Unix. If knowledge in Linux/Unix is required -> 1 , If not mentioned -> 0

SQL. If SQL is required -> 1 , If not mentioned -> 0

NoSQL. If NoSQL is required -> 1 , If not mentioned -> 0

RDBMS. If knowledge in RDBMS is required -> 1 , If not mentioned -> 0

Oracle. If knowledge in Oracle is required -> 1 , If not mentioned -> 0

MySQL. If MYSQL is required -> 1 , If not mentioned -> 0

PHP. If PHP is required -> 1 , If not mentioned -> 0

Flash_Actionscript. If Flash_Actionscript is required -> 1 , If not mentioned -> 0

SPL. If knowledge in SPL is required -> 1 , If not mentioned -> 0

web_design_and_development_tools. If knowledge in Web Design and Development Tools is required -> 1 , If not mentioned -> 0

Wordpress. If Wordpress is required -> 1 , If not mentioned -> 0

AI. If AI is required 1 , If not mentioned 0

Natural_Language_Processing(NLP). If knowledge in NLP is required -> 1 , If not mentioned -> 0

Microsoft_Power_BI. If knowledge in Microsoft Power BI is required -> 1 , If not mentioned -> 0

Google_Analytics. If knowledge in Google Analytics is required -> 1 , If not mentioned -> 0

graphics_and_design_skills. If Graphic and Design Skills are required -> 1 , If not mentioned -> 0

Data_marketing. If Data Marketing abillity is required -> 1 , If not mentioned -> 0

SEO. If knowledge in SEO is required -> 1 , If not mentioned -> 0

Content_Management. If knowledge in Content Management is required -> 1 , If not mentioned -> 0

Tableau. If knowledge in Tableau is required -> 1 , If not mentioned -> 0

D3. If knowledge in D3 is required -> 1 , If not mentioned -> 0

Alteryx. If knowledge in Alteryx is required -> 1 , If not mentioned -> 0

KNIME. If knowledge in KNIME is required -> 1 , If not mentioned -> 0

Spotfire. If knowledge in Spotfire is required -> 1 , If not mentioned -> 0

Spark. If knowledge in Spark is required -> 1 , If not mentioned -> 0

S3. If knowledge in S3 is required -> 1 , If not mentioned -> 0

Redshift. If knowledge in Redshift is required -> 1 , If not mentioned -> 0

DigitalOcean. If knowledge in Digital Ocean is required -> 1 , If not mentioned -> 0

Javascript. If Java Script is required -> 1 , If not mentioned -> 0

Kafka. If knowledge in Kafka is required -> 1 , If not mentioned -> 0

Storm. If knowledge in Storm is required -> 1 , If not mentioned -> 0

Bash. If knowledge in Bash is required -> 1 , If not mentioned -> 0

Hadoop. If knowledge in Hadoop is required -> 1 , If not mentioned -> 0

Data_Pipelines. If knowledge in Data Pipelines is required -> 1 , If not mentioned -> 0

MPP_Platforms. If MPP Platforms is required -> 1 , If not mentioned -> 0

Qlik. If Qlik is required -> 1 , If not mentioned -> 0

Pig. If Pig is required -> 1 , If not mentioned -> 0

Hive. If Hive is required -> 1 , If not mentioned -> 0

Tensorflow. If Tensorflow is required -> 1 , If not mentioned -> 0

Map_Reduce. If Map/Reduce is required -> 1 , If not mentioned -> 0

Impala. If Impala is required -> 1 ,If not mentioned -> 0

Solr. If Sloris required -> 1 , If not mentioned -> 0

Teradata. If Teradata is required -> 1 , If not mentioned -> 0

MongoDB. If MonoDB is required -> 1 , If not mentioned -> 0

Elasticsearch. If Elasticsearch is required -> 1, If not mentioned -> 0

YOLO. If YOLO is required -> 1, If not mentioned -> 0

agile_execution. If agile execution is required -> 1 , If not mentioned -> 0

Data_management. If the knowledge in Data Management is required -> 1 , If not mentioned -> 0

pyspark. If pyspark is required -> 1 , If not mentioned -> 0

Data_mining. If the knowledge in Data Mining is required -> 1 , If not mentioned -> 0

Data_science. If the knowledge in Data Science is required -> 1 , If not mentioned -> 0

Web_Analytic_tools. If the knowledge in Web Analytic tools is required -> 1 , If not mentioned -> 0

IOT. If IOT is required -> 1 , If not mentioned -> 0

Numerical_Analysis. If the knowledge in Numerical Analysis is required -> 1 , If not mentioned -> 0

Economic. If the knowledge in Economic is required -> 1 , If not mentioned -> 0

Finance_Knowledge. If Finance_Knowledge is required -> 1 , If not mentioned -> 0

Investment_Knowledge. If Investment Knowledge is required -> 1 , If not mentioned -> 0

Problem_Solving. If the ability of Problem Solving is required -> 1 , If not mentioned -> 0

Korean_language. If the ability of Korean language is required -> 1 , If not mentioned -> 0

Bash_Linux_Scripting. If Bash Linux Scripting is required -> 1 , If not mentioned -> 0

Team_Handling. If the ability of Team Handling is required -> 1 , If not mentioned -> 0

Debtor_reconcilation. If the ability of Debtor reconciliation is required -> 1 , If not mentioned -> 0

Payroll_management. If the ability of Payroll management is required -> 1 , If not mentioned -> 0

Bayesian. If Bayesian knowledge is required -> 1 , If not mentioned -> 0

Optimization. If Optimization knowledge is required -> 1 ,If not mentioned -> 0

Bahasa_Malaysia. If Bahasa_Malaysia knowledge is required -> 1 ,If not mentioned -> 0

Knowledge_in. Required knowledge to do a particular job , If not mentioned -> NA

City. City where the company is located in , If not mentioned -> NA

Location. Country where the company is located in

Educational_qualifications. Required educational qualifications

Salary. Salary

English_proficiency. English proficiency

URL. URL of the job advertisement

Search_Term. Search Term

Job_Category. Name of the job category

Minimum_Years_of_experience. Minimum years of experience needed for the job , If not mentioned -> NA

Experience. Experience

Experience_Category. Experience category

Job_Country. Job country

Edu_Category. Education category

Minimum_Salary. Minimum salary

Salary_BasisSalary. basis