Data scientists, data analyst, and statistician job advertisements from 2020 to 2023

A dataset with 1172 rows and 109 variables

data(DStidy)

Source

Data collection was done, BSc (Hons)Staistics, University of Sri Jayewardenepura under the statistical consultancy service from 2020 to 2023.

Details

ID. row id

Consultant. Name of the consultant

DateRetrieved. Date of Data Retrieved

DatePublished. Published Date of the Advertisement

Job_title. Name of the job category

Company. Name of the Company

R. If R is required -> 1 ,If not mentioned -> 0

SAS. If SAS is required -> 1 , If not mentioned -> 0

SPSS. If SPSS is required -> 1 , If not mentioned -> 0

Python. If Python is required -> 1 , If not mentioned -> 0

MAtlab. If Matlab is required -> 1 , If not mentioned -> 0

Scala. If Scala is required -> 1 , If not mentioned -> 0

C#. If C# is required -> 1 , If not mentioned -> 0

MS Word. If knowledge in MS Word is required -> 1 , If not mentioned -> 0

Ms Excel. If knowledge in MS Excel is required -> 1 , If not mentioned -> 0

OLE/DB. If knowledge in OLE/DB is required -> 1 , If not mentioned -> 0

Ms Access. If Ms Access is required -> 1 , If not mentioned -> 0

Ms PowerPoint. If knowledge in Ms Powerpoint is required -> 1 , If not mentioned -> 0

Spreadsheets. If knowledge in Spreadsheets is required -> 1 , If not mentioned -> 0

Data_visualization. If knowledge in Data Visualization is required -> 1 , If not mentioned -> 0

Presentation_Skills. If Presentation Skills are required -> 1 , If not mentioned -> 0

Communication. If Communication skills are required -> 1 , If not mentioned -> 0

BigData. If knowledge in Big Data analysis is required -> 1 , If not mentioned -> 0

Data_warehouse. If knowledge in Data Warehouse is required -> 1 , If not mentioned -> 0

cloud_storage. If knowledge in Cloud Storage is required -> 1 , If not mentioned -> 0

Google_Cloud. If knowledge in Google Cloud is required -> 1 , If not mentioned -> 0

AWS. If knowledge in AWS is required -> 1 , If not mentioned -> 0

Machine_Learning. If knowledge in Machine Learning is required -> 1 , If not mentioned -> 0

Deep Learning. If knowledge in Deep Learning is required -> 1 , If not entioned -> 0

Computer_vision. If knowledge in Computer Vision is required -> 1 , If not mentioned -> 0

Java. If Java is required -> 1 , If not mentioned -> 0

C++. If C++ is required -> 1 , If not mentioned -> 0

C. If C is required -> 1 , If not mentioned -> 0

Linux/Unix. If knowledge in Linux/Unix is required -> 1 , If not mentioned -> 0

SQL. If SQL is required -> 1 , If not mentioned -> 0

NoSQL. If NoSQL is required -> 1 , If not mentioned -> 0

RDBMS. If knowledge in RDBMS is required -> 1 , If not mentioned -> 0

Oracle. If knowledge in Oracle is required -> 1 , If not mentioned -> 0

MySQL. If MYSQL is required -> 1 , If not mentioned -> 0

PHP. If PHP is required -> 1 , If not mentioned -> 0

Flash_Actionscript. If knowledge in Flash Action Script is required -> 1 , If not mentioned -> 0

SPL. If knowledge in SPL is required -> 1 , If not mentioned -> 0

web_design_and_development_tools. If knowledge in Web Design and Development Tools is required -> 1 , If not mentioned -> 0

Wordpress. If knowledge in Wordpress is required -> 1 , If not mentioned -> 0

AI. If Artificial Intelligence is required -> 1 , If not mentioned -> 0

Natural_Language_Processing(NLP). If knowledge in NLP is required -> 1 , If not mentioned -> 0

Microsoft Power BI. If knowledge in Microsoft Power BI is required -> 1 , If not mentioned -> 0

Google_Analytics. If knowledge in Google Analytics is required -> 1 , If not mentioned -> 0

graphics_and_design_skills. If Graphic and Design Skills are required -> 1 , If not mentioned -> 0

Data_marketing. If Data Marketing abillity is required -> 1 , If not mentioned -> 0

SEO. If knowledge in SEO is required -> 1 , If not mentioned -> 0

Content_Management. If knowledge in Content Management is required -> 1 , If not mentioned -> 0

Tableau. If knowledge in Tableau is required -> 1 , If not mentioned -> 0

D3. If knowledge in D3 is required -> 1 , If not mentioned -> 0

Alteryx. If knowledge in Alteryx is required -> 1 , If not mentioned -> 0

KNIME. If knowledge in KNIME is required -> 1 , If not mentioned -> 0

Spotfire. If knowledge in Spotfire is required -> 1 , If not mentioned -> 0

Spark. If knowledge in Spark is required -> 1 , If not mentioned -> 0

S3. If knowledge in S3 is required -> 1 , If not mentioned -> 0

Redshift. If knowledge in Redshift is required -> 1 , If not mentioned -> 0

DigitalOcean. If knowledge in Digital Ocean is required -> 1 , If not mentioned -> 0

Javascript. If Java Script is required -> 1 , If not mentioned -> 0

Kafka. If knowledge in Kafka is required -> 1 , If not mentioned -> 0

Storm. If knowledge in Storm is required -> 1 , If not mentioned -> 0

Bash. If knowledge in Bash is required -> 1 , If not mentioned -> 0

Hadoop. If knowledge in Hadoop is required -> 1 , If not mentioned -> 0

Data_Pipelines. If knowledge in Data Pipelines is required -> 1 , If not mentioned -> 0

MPP_Platforms. If MPP Platforms is required ->1,If not mentioned-0

Qlik. If Qlik is required ->1,If not mentioned ->0

Pig. If Pig is required ->1,If not mentioned ->0

Hive. If Hive is required ->1,If not mentioned ->0

Tensorflow. If Tensorflow is required ->1,If not mentioned ->0

Map/Reduce. If Map/Reduce is required ->1,If not mentioned ->0

Impala. If Impala is required ->1,If not mentioned ->0

Solr. If Sloris required ->1,If not mentioned ->0

Teradata. If Teradata is required ->1,If not mentioned ->0

MongoDB. If MonoDB is required ->1,If not mentioned ->0

Elasticsearch. If Elasticsearch is required ->1,If not mentioned ->0

YOLO. If YOLO is required-1 ,If not mentioned-0

agile execution. If agile execution is required->1 ,If not mentioned->0

Data_management. If the knowledge in data management is required->1 ,If not mentioned->0

pyspark. If pyspark is required->1 ,If not mentioned->0

Data_mining. If the knowledge in data mining is required->1 ,If not mentioned->0

Data_science. If the knowledge in data science is required->1 ,If not mentioned->0

Web_Analytic_tools. If the knowledge in Web Analytic tools is required->1 ,If not mentioned->0

IOT. If IOT is required->1 ,If not mentioned->0

Numerical_Analysis. If the knowledge in Numerical Analysis is required->1 ,If not mentioned->0

Economic. If the knowledge in Economic is required->1 ,If not mentioned->0

Finance_Knowledge. If Finance_Knowledge is required->1 ,If not mentioned->0

Investment_Knowledge. If Investment Knowledge is required->1 ,If not mentioned->0

Problem_Solving. If the ability of Problem Solving is required->1 ,If not mentioned->0

Team_Handling. If the ability of Team Handling is required->1 ,If not mentioned->0

Debtor_reconcilation. If the ability of Debtor reconcilation is required->1 ,If not mentioned->0

Payroll_management. If Payroll management is required->1 ,If not mentioned->0

Bayesian. If Bayesian is required->1 ,If not mentioned->0

Optimization. If Optimization knowledge is required-1 ,If not mentioned-0

Knowledge_in. Required knowledge to do a particular job ,If not mentioned->NA

City. City where the company is located in

Educational_qualifications. Required educational qualifications

Salary. Amount of salary

URL. Web address of a particular job advertisement

Search_Term. web search term of a particular job advertisement

Job_Category. Category of the job (i.e. "Data Science","Data Analyst" etc.)

Team_Handling. If the ability of Team Handling is required-1 ,If not mentioned-0

Debtor_reconcilation. If the ability of Debtor reconciliation is required-1 ,If not mentioned-0

Payroll_management. If the ability of Payroll management is required-1 ,If not mentioned-0

Bayesian. If Bayesian knowledge is required-1 ,If not mentioned-0

Bahasa_Malaysia. If Bahasa Malaysia is required-1 ,If not mentioned-0

English_proficiency. If English proficiency is required-1 ,If not mentioned-0

Experience_Category. Number of years of experience in binned into categories

Location. Location

Payment Frequency. Payment frequency

BSc_needed. If BSc is required-1 ,If not mentioned-0

MSc_needed. If MSc is required-1 ,If not mentioned-0

PhD_needed. If PhD is required-1 ,If not mentioned-0

English Needed. If English is required-1 ,If not mentioned-0

year. Survey year